Display Market balance and price on mobile
